requirements.txt
blessed==1.20.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
cycler==0.11.0
einops==0.6.1
gpustat==1.1
idna==3.4
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.3.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.6
nvidia-ml-py==12.535.77
opencv-python==4.5.5.62
packaging==23.1
pandas==1.1.5
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboardX==2.6.1
thop==0.1.1-2209072238
threadpoolctl==3.1.0
torchinfo==1.8.0
tqdm==4.61.2
typing_extensions==4.7.1
urllib3==1.26.14
wcwidth==0.2.6
xlrd==1.2.0
